Fix test failure. I have no idea why this couldn't be in database.mts.
