+++ server/src/services/vatsimATIS.mts
@@ -2,7 +2,7 @@ import _ from "lodash";
 import { IVatsimATIS, IVatsimData } from "../interfaces/IVatsimData.mjs";
 import mainLogger from "../logger.mjs";
 import { VatsimATISDocument, VatsimATISModel } from "../models/VatsimATIS.mjs";
-import { logMongoBulkErrors } from "../database.mjs";
+import { logMongoBulkErrors } from "../utils.mjs";
 
 const logger = mainLogger.child({ service: "vatsimATIS" });
